test(migrations): split MigrationTest mega-file into focused suites (#902)
The 22 test methods that lived in tests/MigrationTest.php now live in
seven focused suites under tests/Migration/, each named for the migration
era it covers:

  - SqliteOnlyClosuresTest    — VRF rebuild + PRAGMA foreign_keys restore (6)
  - EngineParityTest          — apply_migrations() idempotency + fresh-schema
                                replay (2)
  - SettingsCascadeTest       — v2.6.0 settings seed + v3.13.0 tenant_id (2)
  - MfaTest                   — v3.15.0 passkeys, v3.16.0 preferred MFA (2)
  - BackupTest                — v3.17.0+ destinations/schedules, v3.21.0
                                backup_runs row-copy, schedule-unique dedup (3)
  - IpStorageTest             — v2.9.0 BLOB-affinity flip + idempotency (2)
  - MiscTest                  — v3.2.0 devices, v3.5.0 custom-fields,
                                v3.6.0 TOTP/lockout scaffold (5)

Total: 22 tests / 146 assertions — byte-equivalent to the pre-split
baseline (verified with --testdox).

Shared makeDb() / makePreVrfDb() fixtures live in tests/Migration/Base.php
under the Tests\Migration namespace (PSR-4 wired via autoload-dev in
composer.json — same mapping used for #903's Tests\Helpers\InMemoryDb).

tests/MigrationTest.php stays as a sanity stub that just asserts the
seven split files plus Base.php exist — preserves CI/contributor muscle
memory pointing at the old class name.
